requirements.txt
discord.py==2.4.0
supabase==2.4.3
aiohttp==3.9.5
matplotlib==3.8.4
python-dotenv==1.0.1
flask==3.0.3
requests==2.32.3
